requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.13
exceptiongroup==1.2.1
executing==2.0.1
gunicorn==22.0.0
idna==3.7
ipython==8.18.1
ipython-genutils==0.2.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
djangorestframework~=3.15.2